fix(appraisal): editing a building with a fractional floor count no longer 400s
NumberOfFloors is decimal(5,2) in the database and decimal? on the domain
entity, but several request/command records were still int? — left behind
when the domain was widened in ff2cea38. Posting 2.5 failed at model
binding, surfacing to the user as "Invalid value at '$.numberOfFloors'.
Please check the data type and format.", so a building that could be
CREATED with 2.5 floors could never be EDITED.

Also closes a silent-truncation path: UpdateCondoPropertyRequest was
decimal? while its command was int?, and the endpoint maps between them
with Mapster, so the fractional part was quietly floored on its way to a
decimal column — no error, just a wrong number in the database.

Widening only; no handler, domain, EF configuration or migration changes
are needed.

+++ Modules/Appraisal/Appraisal/Application/Features/Appraisals/CreateCondoProperty/CreateCondoPropertyCommand.cs
@@ -62,7 +62,7 @@ public record CreateCondoPropertyCommand(
     string? DecorationTypeOther = null,
     int? BuildingAge = null,
     int? ConstructionYear = null,
-    int? NumberOfFloors = null,
+    decimal? NumberOfFloors = null,
     string? BuildingFormType = null,
     string? ConstructionMaterialType = null,
     // Layout & Materials
